+++ investor/adapters/repositories/asset/in_memory/repository.go
@@ -6,7 +6,7 @@ import (
 )
 
 type InMemoryAssetRepository struct {
-	payments map[string]assetEntity.Asset
+	assets map[string]assetEntity.Asset
 }
 
 type AssetAlreadyExitsError struct {
@@ -17,16 +17,16 @@ func (e AssetAlreadyExitsError) Error() string {
 	return fmt.Sprintf("payment with id %s already exists", e.AssetId)
 }
 
-func (storage *InMemoryAssetRepository) Create(asset assetEntity.Asset) (err error) {
-	_, idExists := storage.payments[asset.Id]
+func (repository *InMemoryAssetRepository) Create(asset assetEntity.Asset) (err error) {
+	_, idExists := repository.assets[asset.Id]
 	if idExists {
 		err = AssetAlreadyExitsError{AssetId: asset.Id}
 	} else {
-		storage.payments[asset.Id] = asset
+		repository.assets[asset.Id] = asset
 	}
 	return
 }
 
 func NewInMemoryAssetRepository() *InMemoryAssetRepository {
-	return &InMemoryAssetRepository{payments: make(map[string]assetEntity.Asset)}
+	return &InMemoryAssetRepository{assets: make(map[string]assetEntity.Asset)}
 }
