+++ investor/adapters/repositories/payment/json/json_reader.go
